requirements.txt
pandas
numpy<1.24
tensorflow<2.11
matplotlib
pillow
seaborn
plotly
scikit-learn
opendatasets
IPython
